pyproject.toml
[project]
name = "question-answer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8",
    "django-rest-framework>=0.1.0",
    "drf-spectacular>=0.29.0",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-django>=4.11.1",
]
